pyproject.toml
[project]
name = "stack"
version = "0.1.0"
description = "Full Stack ML"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]",
    "hopsworks[python]==4.2.2",
    "hsfs>=3.7.9",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "pandas>=2.1.4",
    "polars>=1.32.3",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stack"]
